requirements.txt
appium-python-client
selenium
PyYAML
HTMLTestRunner-Python3
pytest
pytest-html
